Simulator spi bsp (#77)
* Makefile deletes *.out files when make clean is called

* Adds file reads in BSP_SPI read

* Adds skeleton to find the command.

* Adds PEC calculations and half of the command ltc6811 command codes

* Adds file reads in BSP_SPI read

* Adds skeleton to find the command.

* Adds PEC calculations and half of the command ltc6811 command codes

* Adds Test files for SPI and LTC6811 and finishes the command codes for the LTC6811

* Adds function to format the packets the LTC6811 would usually transmit inside BSP_SPI.c

* Adds LTC6811 command handler into BSP_SPI to simulate the BSP. Adds helper functions to easily copy and extract data. Formats the doc for BSP_SPI to make everything more readable.

* Subtracts 1 from starting indices in for loop of BSP_SPI CmdHandler function. The LTC6811 sends the data of the last LTC6811 in daisy chain first so need to populate the data struct in reverse but forgot to start at NUM_MINIONS-1

* Formatting of the voltage data for the LTC6811 works.

* Fixes warnings in LTC681x.c. Also adds untested functionality of selecting and returning the temperature sensor data in BSP_SPI.c

* Creates temperature test to test out BSP_SPI with temperature sensors

* Fix Lights simulator bug

* Generates voltage and temperature data within ranges specified by the state and mode parameters

* SPI simulation works with the battery model

* Documentation & read function

* Update display to show module connection, voltage, and temperature

* Fix Lights.csv not existing bug on first run

* User can disconnect and reconnect wires while simulating the real battery

* Change precision of Voltage to 0.0001 and Temperature to 0.01

* Update LED display to include colors; Make lights default to off on initialization

* Adds file reads in BSP_SPI read

* Adds skeleton to find the command.

* Adds PEC calculations and half of the command ltc6811 command codes

* Adds file reads in BSP_SPI read

* Adds PEC calculations and half of the command ltc6811 command codes

* Adds Test files for SPI and LTC6811 and finishes the command codes for the LTC6811

* Adds function to format the packets the LTC6811 would usually transmit inside BSP_SPI.c

* Adds LTC6811 command handler into BSP_SPI to simulate the BSP. Adds helper functions to easily copy and extract data. Formats the doc for BSP_SPI to make everything more readable.

* Subtracts 1 from starting indices in for loop of BSP_SPI CmdHandler function. The LTC6811 sends the data of the last LTC6811 in daisy chain first so need to populate the data struct in reverse but forgot to start at NUM_MINIONS-1

* Formatting of the voltage data for the LTC6811 works.

* Fixes warnings in LTC681x.c. Also adds untested functionality of selecting and returning the temperature sensor data in BSP_SPI.c

* Creates temperature test to test out BSP_SPI with temperature sensors

* Formatting of temperature sensor data working. Need to get add code to extract data from csv files now.

* Reads csv file for voltage correctly. Temperature formatting incorrect though.

* Temperature is being read correctly.

* Changes temperature format to have 3 dec places and adds '>>' to indicate whenever the simulate is waiting for an input

* Adds print statement to make the terminal window bigger if simulator catches an exception from addwstr()

* Adds skeleton and test code to support open wire checks

* Adds semaphores for handling open wire voltages for open wire check in BSP_SPI

* Extends ===== border in simulation for Modules section

* Open wire check working with simulator. There are still issues to be resolved but better suited to be handled in a different branch. Issues that must be fixed will be issued on github. The BSP_SPI itself is working. Battery Balancing is NOT supported though.

* Forgot to call fclose after finishing with file IO operations

* Deletes erreneous white space

* Fixes compiler warnings

* Adds checks to BSP Init functions to make sure files exist

* Accidentally used negative logic for open wires. Fixed it

* Changes max temperature to be 72C for normal so bps won't fault during normal operations. There is a floating point error during the conversion process so 73.000C for the simulator will read 73.000 +- 0.05 error for the BPS

* Adds locks to SPI.csv reads and writes

* Program exits if fopen failes

+++ BeVolt/Apps/Inc/Temperature.h
@@ -87,7 +87,7 @@ uint8_t *Temperature_GetModulesInDanger(void);
  * @param index of sensor (0-indexed based)
  * @return temperature of the battery module at specified index
  */
-int16_t Temperature_GetSingleTempSensor(uint8_t board, uint8_t sensorIdx);
+int32_t Temperature_GetSingleTempSensor(uint8_t board, uint8_t sensorIdx);
 
 /** Temperature_GetModuleTemperature
  * Gets the avg temperature of a certain battery module in the battery pack. Since there
@@ -96,13 +96,13 @@ int16_t Temperature_GetSingleTempSensor(uint8_t board, uint8_t sensorIdx);
  * @param index of module (0-indexed based)
  * @return temperature of the battery module at specified index
  */
-int16_t Temperature_GetModuleTemperature(uint8_t moduleIdx);
+int32_t Temperature_GetModuleTemperature(uint8_t moduleIdx);
 
 /** Temperature_GetTotalPackAvgTemperature
  * Gets the average temperature of the whole battery pack
  * @return average temperature of battery pack
  */
-int16_t Temperature_GetTotalPackAvgTemperature(void);
+int32_t Temperature_GetTotalPackAvgTemperature(void);
 
 /** Temperature_SampleADC
  * Starts ADC conversion on GPIO1 on LTC6811's auxiliary registers on all boards
